refactor: WriterAvatar를 comment 엔티티로 추출, getMe 임포트 경로 수정 (#151)
