+++ src/entities/comment/index.ts
@@ -3,3 +3,5 @@ export type { Writer, Comment, CommentListResponse } from "./model/schema";
 
 export { useRecentComments } from "./api/useRecentComments";
 export { useEpigramComments } from "./api/useEpigramComments";
+
+export { WriterAvatar } from "./ui/WriterAvatar";
